+++ src/main/java/com/audiomaster/audio/processor/processor.java
@@ -1,0 +1,5 @@
+package com.audiomaster.audio.processor;
+
+public interface processor {
+
+}
