+++ src/main/java/com/audiomaster/audio/type/EqType.java
@@ -1,4 +1,4 @@
-package com.audiomaster.plugin.type;
+package com.audiomaster.audio.type;
 
 public enum EqType {
     LOWSHELF, HIGHSHELF, BAND
