--- src/main/java/com/audiomaster/plugin/processor.java
@@ -1,8 +1,0 @@
-package com.audiomaster.plugin;
-
-import java.util.List;
-
-public interface processor {
-
-    List<String> toElement();
-}
